# wfzo-api-main/wfzo-api-main/api-spec.yml
openapi: 3.0.4
info:
  title: WFZO Website Management Platform API
  description: >-
    REST APIs for the World Free Zones Organization (WFZO) web platform.
    Scopes
     - Auth: Microsoft Entra (External ID) Bearer JWT
     - Content: read-through to Strapi CMS
     - Documents: secure PDF delivery via short-lived Azure Blob SAS
     - Events: Zoom registration with per-membership seat quotas
     - Learning: LMS integration with Moodle
     - Community: Discourse integrations (DMs, topics, embeds)
     - CMS APIs
  version: 1.0.0
servers:
  - url: /wfzo/api/v1
...
paths:
...
  /mailerlite/subscribers:
    post:
      tags:
        - MailerLite
      summary: Create or update a subscriber
      description: Add a new subscriber or update an existing one in MailerLite (Legacy v2 API). The default group (MAILERLITE_GROUP_ID) will be automatically added to all subscribers after creation.
      operationId: upsertSubscriber
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertSubscriberDto"
            examples:
              newSubscriber:
                summary: Create a new subscriber (default group will be auto-added)
                value:
                  email: "john.doe@example.com"
                  fields:
                    name: "John Doe"
                    company: "Example Corp"
                  groups: ["group_id_1", "group_id_2"]
              updateSubscriber:
                summary: Update existing subscriber
                value:
                  email: "john.doe@example.com"
                  fields:
                    name: "John Doe"
                    company: "Updated Corp"
                    phone: "+1234567890"
      responses:
        "200":
          description: Subscriber created or updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
              examples:
                success:
                  value:
                    code: "SUCCESS"
                    type: "SUCCESS"
                    message: "Subscriber upserted successfully"
                    data:
                      id: "123456789"
                      email: "john.doe@example.com"
                      status: "active"
                      fields:
                        name: "John Doe"
                        company: "Example Corp"
                      groups: ["113129360"]
        "400":
          description: Invalid input
        default:
          description: Unexpected error
  /mailerlite/subscribers/{email}:
    get:
      tags:
        - MailerLite
      summary: Get subscriber by email
      description: Retrieve subscriber details from MailerLite
      operationId: getSubscriber
      parameters:
        - name: email
          in: path
          required: true
          schema:
            type: string
            format: email
          description: Subscriber email address
      responses:
        "200":
          description: Subscriber found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
              examples:
                subscriberFound:
                  value:
                    code: "SUCCESS"
                    type: "SUCCESS"
                    message: "Subscriber retrieved successfully"
                    data:
                      id: "123456789"
                      email: "john.doe@example.com"
                      status: "active"
                      fields:
                        name: "John Doe"
                        company: "Example Corp"
                        phone: "+1234567890"
                      groups: ["113129360"]
                      subscribed_at: "2023-01-15T10:30:00Z"
        "404":
          description: Subscriber not found
        default:
          description: Unexpected error
    delete:
      tags:
        - MailerLite
      summary: Delete a subscriber
      description: Remove a subscriber from MailerLite
      operationId: deleteSubscriber
      parameters:
        - name: email
          in: path
          required: true
          schema:
            type: string
            format: email
          description: Subscriber email address
      responses:
        "200":
          description: Subscriber deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
              examples:
                deleted:
                  value:
                    code: "SUCCESS"
                    type: "SUCCESS"
                    message: "Subscriber deleted successfully"
                    data:
                      success: true
        "404":
          description: Subscriber not found
        default:
          description: Unexpected error
components:
...
  schemas:
...
    ApiResponse:
      type: object
      description: Standard API response envelope for status reporting
      properties:
        code:
          type: integer
          format: int32
        type:
          type: string
        message:
          type: string
      xml:
        name: "##default"
...
    UpsertSubscriberDto:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
          description: Subscriber email address
        fields:
          type: object
          additionalProperties:
            type: string
          description: Custom fields for the subscriber
        groups:
          type: array
          items:
            type: string
          description: List of group IDs to add the subscriber to
  securitySchemes:
    petstore_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://petstore3.swagger.io/oauth/authorize
          scopes:
            write:pets: modify pets in your account
            read:pets: read your pets
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    api_key:
      type: apiKey
      name: api_key
      in: header